refactor: create a lib directory #1
